Move FlattenGrid from aic_nd.py to utils_nd.py
